# openapi_gpt_static.json
{
  "openapi": "3.1.0",
  "info": { "title": "SchedulerGPT API", "version": "gpt-static-0.3.1" },
  "servers": [{ "url": "https://scheduler-gpt-api.onrender.com" }],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": { "type": "apiKey", "in": "header", "name": "X-API-Key" }
    },
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": { "ok": { "type": "boolean" } },
        "required": ["ok"]
      },
      "JobPoolResponse": {
        "type": "object",
        "properties": {
          "count": { "type": "integer" },
          "jobs": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        },
        "required": ["count", "jobs"]
      },
      "TechniciansResponse": {
        "type": "object",
        "properties": {
          "count": { "type": "integer" },
          "technicians": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        },
        "required": ["count", "technicians"]
      },
      "ScheduleExistingResponse": {
        "type": "object",
        "properties": {
          "count": { "type": "integer" },
          "rows": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        },
        "required": ["count", "rows"]
      },
      "ValidateRequest": {
        "type": "object",
        "properties": {
          "work_order": { "type": "integer" },
          "technician_id": { "type": "integer" },
          "date": { "type": "string", "format": "date" },
          "start_time": { "type": "string", "nullable": true, "description": "HH:MM" },
          "end_time": { "type": "string", "nullable": true, "description": "HH:MM" }
        },
        "required": ["work_order", "technician_id", "date"]
      },
      "ValidateResponse": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "errors": { "type": "array", "items": { "type": "string" } },
          "warnings": { "type": "array", "items": { "type": "string" } },
          "metrics": { "type": "object", "additionalProperties": true }
        },
        "required": ["ok", "errors", "warnings", "metrics"]
      },
      "AssignJobRequest": {
        "type": "object",
        "properties": {
          "work_order": { "type": "integer" },
          "technician_id": { "type": "integer" },
          "date": { "type": "string", "format": "date" },
          "start_time": { "type": "string", "nullable": true },
          "end_time": { "type": "string", "nullable": true },
          "is_night": { "type": "boolean", "nullable": true },
          "region": { "type": "string", "nullable": true },
          "cluster_id": { "type": "integer", "nullable": true }
        },
        "required": ["work_order", "technician_id", "date"]
      },
      "AssignJobResponse": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "assigned": { "type": "object", "additionalProperties": true }
        },
        "required": ["ok", "assigned"]
      }
    }
  },
  "security": [{ "ApiKeyAuth": [] }],
  "paths": {
    "/health": {
      "get": {
        "operationId": "health",
        "summary": "Service health check",
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HealthResponse" } } }
          }
        }
      }
    },
    "/jobs/pool": {
      "get": {
        "operationId": "jobsPool",
        "summary": "List jobs in job_pool by due date window",
        "parameters": [
          { "name": "start", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } },
          { "name": "end", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } },
          { "name": "states", "in": "query", "required": false, "schema": { "type": "string" }, "description": "CSV like CO,UT,AZ" },
          { "name": "statuses", "in": "query", "required": false, "schema": { "type": "string" }, "description": "CSV. Use '*' or omit for no filter." },
          { "name": "is_night", "in": "query", "required": false, "schema": { "type": "boolean" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 5000, "default": 500 } }
        ],
        "responses": {
          "200": {
            "description": "Jobs",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobPoolResponse" } } }
          }
        }
      }
    },
    "/technicians": {
      "get": {
        "operationId": "technicians",
        "summary": "List technicians and capacities",
        "parameters": [
          { "name": "active_only", "in": "query", "required": false, "schema": { "type": "boolean", "default": true } },
          { "name": "night_eligible", "in": "query", "required": false, "schema": { "type": "boolean" } },
          { "name": "region", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 5000, "default": 500 } }
        ],
        "responses": {
          "200": {
            "description": "Technicians",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TechniciansResponse" } } }
          }
        }
      }
    },
    "/schedule/existing": {
      "get": {
        "operationId": "scheduleExisting",
        "summary": "Get existing scheduled jobs for a window",
        "parameters": [
          { "name": "start", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } },
          { "name": "end", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } },
          { "name": "technician_ids", "in": "query", "required": false, "schema": { "type": "string" }, "description": "CSV numeric IDs" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 5000, "default": 1000 } }
        ],
        "responses": {
          "200": {
            "description": "Scheduled rows",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScheduleExistingResponse" } } }
          }
        }
      }
    },
    "/schedule/validate": {
      "post": {
        "operationId": "scheduleValidate",
        "summary": "Validate a single assignment against rules",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidateRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Validation result",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidateResponse" } } }
          }
        }
      }
    },
    "/jobs/assign": {
      "post": {
        "operationId": "assignJob",
        "summary": "Commit a single job to a technician and date",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AssignJobRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Assignment result",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AssignJobResponse" } } }
          }
        }
      }
    }
  }
}